requirements.txt
# Core dependencies
pyspark==3.5.0
fastapi==0.104.1
uvicorn[standard]==0.24.0.post1
sqlalchemy==2.0.23
alembic==1.12.1
celery[redis]==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
boto3==1.29.7

# API dependencies
python-multipart==0.0.6
httpx==0.25.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2

# Monitoring and logging
prometheus-client==0.19.0
structlog==23.2.0
python-json-logger==2.0.7

# Data processing
pandas==2.1.3
numpy==1.26.2
pyarrow==14.0.1
dask==2023.11.0
scikit-learn==1.3.2

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
behave==1.2.6
black==23.11.0
flake8==6.1.0
mypy==1.7.1
isort==5.12.0
pre-commit==3.5.0

# Type stubs
types-redis==4.6.0.20240106
types-requests==2.31.0.10
types-python-dateutil==2.8.19.14

# Additional utilities
click==8.1.7
pyyaml==6.0.1
python-dotenv==1.0.0
tenacity==8.2.3
croniter==2.0.1